change server logic
